+++ src/main/java/calculator/view/FormulaInput.java
@@ -1,4 +1,4 @@
-package calculator.ui;
+package calculator.view;
 
 import calculator.domain.Formula;
 import calculator.util.Console;
