+++ src/main/java/racingcar/view/RacingCarInput.java
@@ -1,4 +1,4 @@
-package racingcar.ui;
+package racingcar.view;
 
 import java.util.Arrays;
 import java.util.List;
